Move interfaces de Clarice para clarice/interfaces
